test-configs.yaml: Add a machine type for Juno with ACPI firmware
The standard LAVA interpretation of Juno includes support for rewriting
the firmware as part of the test job but not all labs will want to do
that since it makes jobs take longer to run and be more hassle to
specify. Add a separate device type for Juno r0 systems with fixed ACPI
firmware, treating it as a generic SBSA device booting using grub.

Signed-off-by: Mark Brown <broonie@kernel.org>